+++ api/daos/gorms/EliminatoriasGorm.go
@@ -3,9 +3,9 @@ package gorms
 import "database/sql"
 
 type EliminatoriasGorm struct {
-	IDCampeonato   int           `gorm:"column:id_campeonato"`
-	IDEliminatoria int           `gorm:"column:id_eliminatoria;primary_key"`
-	IDPartido      int           `gorm:"column:id_partido"`
+	IDCampeonato   int64           `gorm:"column:id_campeonato"`
+	IDEliminatoria int64           `gorm:"column:id_eliminatoria;primary_key"`
+	IDPartido      int64           `gorm:"column:id_partido"`
 	NroLlave       sql.NullInt64 `gorm:"column:nro_llave"`
 }
