+++ api/daos/gorms/JugadoresGorm.go
@@ -1,8 +1,8 @@
 package gorms
 
 type JugadoresGorm struct {
-	IDJugador int `gorm:"column:id_jugador;primary_key"`
-	IDPersona int `gorm:"column:id_persona"`
+	IDJugador int64 `gorm:"column:id_jugador;primary_key"`
+	IDPersona int64 `gorm:"column:id_persona"`
 }
 
 // TableName sets the insert table name for this struct type
