+++ api/daos/gorms/NotificacionesGorm.go
@@ -4,7 +4,7 @@ import "database/sql"
 
 type NotificacionesGorm struct {
 	IDGrupo        sql.NullInt64 `gorm:"column:id_grupo"`
-	IDNotificacion int           `gorm:"column:id_notificacion;primary_key"`
+	IDNotificacion int64           `gorm:"column:id_notificacion;primary_key"`
 	Texto          string        `gorm:"column:texto"`
 	Titulo         string        `gorm:"column:titulo"`
 }
