+++ api/daos/gorms/ZonasEquiposGorm.go
@@ -1,8 +1,8 @@
 package gorms
 
 type ZonasEquiposGorm struct {
-	IDEquipo int `gorm:"column:id_equipo;primary_key"`
-	IDZona   int `gorm:"column:id_zona;primary_key"`
+	IDEquipo int64 `gorm:"column:id_equipo;primary_key"`
+	IDZona   int64 `gorm:"column:id_zona;primary_key"`
 }
 
 // TableName sets the insert table name for this struct type
