+++ api/daos/gorms/ZonasGorm.go
@@ -3,8 +3,8 @@ package gorms
 import "database/sql"
 
 type ZonasGorm struct {
-	IDCampeonato int            `gorm:"column:id_campeonato"`
-	IDZona       int            `gorm:"column:id_zona;primary_key"`
+	IDCampeonato int64            `gorm:"column:id_campeonato"`
+	IDZona       int64            `gorm:"column:id_zona;primary_key"`
 	Nombre       sql.NullString `gorm:"column:nombre"`
 }
